Extract condition to its own file
