+++ packages/alfa-media/src/media.ts
@@ -6,10 +6,10 @@ import { Serializable } from "@siteimprove/alfa-json";
 import { Option, None } from "@siteimprove/alfa-option";
 import { Parser } from "@siteimprove/alfa-parser";
 import { Predicate } from "@siteimprove/alfa-predicate";
-import { Slice } from "@siteimprove/alfa-slice";
 
 import * as json from "@siteimprove/alfa-json";
 
+import { Condition } from "./condition";
 import { Feature } from "./feature/feature";
 import { Type } from "./type";
 
@@ -21,8 +21,6 @@ const {
   left,
   right,
   delimited,
-  oneOrMore,
-  zeroOrMore,
   separatedList,
   eof,
 } = Parser;
@@ -47,11 +45,6 @@ export namespace Media {
     map(Token.parseIdent("not"), () => Modifier.Not)
   );
 
-  export enum Combinator {
-    And = "and",
-    Or = "or",
-  }
-
   export enum Comparator {
     GreaterThan = ">",
     GreaterThanEqual = ">=",
@@ -62,254 +55,25 @@ export namespace Media {
   export const { of: type, isType } = Type;
 
   /**
-   * @see https://drafts.csswg.org/mediaqueries/#media-condition
-   */
-  export class Condition implements Queryable<Condition.JSON> {
-    public static of(
-      combinator: Combinator,
-      left: Feature | Condition | Negation,
-      right: Feature | Condition | Negation
-    ): Condition {
-      return new Condition(combinator, left, right);
-    }
-
-    private readonly _combinator: Combinator;
-    private readonly _left: Feature | Condition | Negation;
-    private readonly _right: Feature | Condition | Negation;
-
-    private constructor(
-      operator: Combinator,
-      left: Feature | Condition | Negation,
-      right: Feature | Condition | Negation
-    ) {
-      this._combinator = operator;
-      this._left = left;
-      this._right = right;
-    }
-
-    public get combinator(): Combinator {
-      return this._combinator;
-    }
-
-    public get left(): Feature | Condition | Negation {
-      return this._left;
-    }
-
-    public get right(): Feature | Condition | Negation {
-      return this._right;
-    }
-
-    public matches(device: Device): boolean {
-      switch (this._combinator) {
-        case Combinator.And:
-          return this._left.matches(device) && this._right.matches(device);
-
-        case Combinator.Or:
-          return this._left.matches(device) || this._right.matches(device);
-      }
-    }
-
-    public equals(value: unknown): value is this {
-      return (
-        value instanceof Condition &&
-        value._combinator === this._combinator &&
-        value._left.equals(this._left) &&
-        value._right.equals(this._right)
-      );
-    }
-
-    public toJSON(): Condition.JSON {
-      return {
-        type: "condition",
-        combinator: this._combinator,
-        left: this._left.toJSON(),
-        right: this._right.toJSON(),
-      };
-    }
-
-    public toString(): string {
-      return `(${this._left}) ${this._combinator} (${this._right})`;
-    }
-  }
-
-  export namespace Condition {
-    export interface JSON {
-      [key: string]: json.JSON;
-      type: "condition";
-      combinator: string;
-      left: Feature.JSON | Condition.JSON | Negation.JSON;
-      right: Feature.JSON | Condition.JSON | Negation.JSON;
-    }
-  }
-
-  export function isCondition(value: unknown): value is Condition {
-    return value instanceof Condition;
-  }
-
-  export class Negation implements Queryable<Negation.JSON> {
-    public static of(condition: Feature | Condition | Negation): Negation {
-      return new Negation(condition);
-    }
-
-    private readonly _condition: Feature | Condition | Negation;
-
-    private constructor(condition: Feature | Condition | Negation) {
-      this._condition = condition;
-    }
-
-    public get condition(): Feature | Condition | Negation {
-      return this._condition;
-    }
-
-    public matches(device: Device): boolean {
-      return !this._condition.matches(device);
-    }
-
-    public equals(value: unknown): value is this {
-      return (
-        value instanceof Negation && value._condition.equals(this._condition)
-      );
-    }
-
-    public toJSON(): Negation.JSON {
-      return {
-        type: "negation",
-        condition: this._condition.toJSON(),
-      };
-    }
-
-    public toString(): string {
-      return `not (${this._condition})`;
-    }
-  }
-
-  export namespace Negation {
-    export interface JSON {
-      [key: string]: json.JSON;
-      type: "negation";
-      condition: Feature.JSON | Condition.JSON | Negation.JSON;
-    }
-  }
-
-  export function isNegation(value: unknown): value is Negation {
-    return value instanceof Negation;
-  }
-
-  /**
-   * @remarks
-   * The condition parser is forward-declared as it is needed within its
-   * subparsers.
-   */
-  let parseCondition: Parser<
-    Slice<Token>,
-    Feature | Condition | Negation,
-    string
-  >;
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-in-parens
-   */
-  const parseInParens = either(
-    delimited(
-      Token.parseOpenParenthesis,
-      delimited(option(Token.parseWhitespace), (input) =>
-        parseCondition(input)
-      ),
-      Token.parseCloseParenthesis
-    ),
-    Feature.parse
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-not
-   */
-  const parseNot = map(
-    right(
-      delimited(option(Token.parseWhitespace), Token.parseIdent("not")),
-      parseInParens
-    ),
-    (condition) => Negation.of(condition)
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-and
-   */
-  const parseAnd = right(
-    delimited(option(Token.parseWhitespace), Token.parseIdent("and")),
-    parseInParens
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-or
-   */
-  const parseOr = right(
-    delimited(option(Token.parseWhitespace), Token.parseIdent("or")),
-    parseInParens
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-condition
-   */
-  parseCondition = either(
-    parseNot,
-    either(
-      map(
-        pair(
-          parseInParens,
-          either(
-            map(
-              oneOrMore(parseAnd),
-              (queries) => [Combinator.And, queries] as const
-            ),
-            map(
-              oneOrMore(parseOr),
-              (queries) => [Combinator.Or, queries] as const
-            )
-          )
-        ),
-        ([left, [combinator, right]]) =>
-          Iterable.reduce(
-            right,
-            (left, right) => Condition.of(combinator, left, right),
-            left
-          )
-      ),
-      parseInParens
-    )
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-condition-without-or
-   */
-  const parseConditionWithoutOr = either(
-    parseNot,
-    map(pair(parseInParens, zeroOrMore(parseAnd)), ([left, right]) =>
-      [left, ...right].reduce((left, right) =>
-        Condition.of(Combinator.And, left, right)
-      )
-    )
-  );
-
-  /**
    * @see https://drafts.csswg.org/mediaqueries/#media-query
    */
   export class Query implements Queryable<Query.JSON> {
     public static of(
       modifier: Option<Modifier>,
       mediaType: Option<Type>,
-      condition: Option<Feature | Condition | Negation>
+      condition: Option<Condition.Condition>
     ): Query {
       return new Query(modifier, mediaType, condition);
     }
 
     private readonly _modifier: Option<Modifier>;
     private readonly _mediaType: Option<Type>;
-    private readonly _condition: Option<Feature | Condition | Negation>;
+    private readonly _condition: Option<Condition.Condition>;
 
     private constructor(
       modifier: Option<Modifier>,
       mediaType: Option<Type>,
-      condition: Option<Feature | Condition | Negation>
+      condition: Option<Condition.Condition>
     ) {
       this._modifier = modifier;
       this._mediaType = mediaType;
@@ -324,7 +88,7 @@ export namespace Media {
       return this._mediaType;
     }
 
-    public get condition(): Option<Feature | Condition | Negation> {
+    public get condition(): Option<Condition.Condition> {
       return this._condition;
     }
 
@@ -380,7 +144,7 @@ export namespace Media {
       [key: string]: json.JSON;
       modifier: string | null;
       mediaType: Type.JSON | null;
-      condition: Feature.JSON | Condition.JSON | Negation.JSON | null;
+      condition: Condition.JSON | null;
       type: "query";
     }
   }
@@ -389,7 +153,7 @@ export namespace Media {
    * @see https://drafts.csswg.org/mediaqueries/#typedef-media-query
    */
   const parseQuery = either(
-    map(parseCondition, (condition) =>
+    map(Condition.parse, (condition) =>
       Query.of(None, None, Option.of(condition))
     ),
     map(
@@ -401,7 +165,7 @@ export namespace Media {
         option(
           right(
             delimited(option(Token.parseWhitespace), Token.parseIdent("and")),
-            parseConditionWithoutOr
+            Condition.parseWithoutOr
           )
         )
       ),
